eSPI: move espi_utils.h to include directory
Move the espi_utils.h header file to include directory to allow
it to be included from the tests.
