+++ drivers/espi/host_subs_npcx.c
@@ -113,13 +113,13 @@
  */
 
 #include <assert.h>
-#include <zephyr/drivers/espi.h>
+#include <zephyr/drivers/espi/espi.h>
+#include <zephyr/drivers/espi/espi_utils.h>
 #include <zephyr/drivers/gpio.h>
 #include <zephyr/drivers/clock_control.h>
 #include <zephyr/drivers/pinctrl.h>
 #include <zephyr/kernel.h>
 #include <soc.h>
-#include "espi_utils.h"
 #include "soc_host.h"
 #include "soc_espi.h"
 #include "soc_miwu.h"
