+++ include/zephyr/drivers/emul.h
@@ -23,9 +23,9 @@ struct emul;
 /* #includes required after forward-declaration of struct emul later defined in this header. */
 #include <zephyr/device.h>
 #include <zephyr/devicetree.h>
-#include <zephyr/drivers/espi_emul.h>
 #include <zephyr/drivers/i2c_emul.h>
 #include <zephyr/drivers/spi_emul.h>
+#include <zephyr/drivers/espi/espi_emul.h>
 
 /**
  * The types of supported buses.
