+++ samples/drivers/espi/src/espi_oob_handler.c
@@ -7,7 +7,7 @@
 #include <zephyr/device.h>
 #include <soc.h>
 #include <zephyr/drivers/gpio.h>
-#include <zephyr/drivers/espi.h>
+#include <zephyr/drivers/espi/espi.h>
 #include <zephyr/logging/log_ctrl.h>
 #include <zephyr/logging/log.h>
 #include "espi_oob_handler.h"
