+++ soc/arm/nuvoton_npcx/common/power.c
@@ -48,7 +48,7 @@
 #include <zephyr/arch/arm/aarch32/cortex_m/cmsis.h>
 #include <zephyr/kernel.h>
 #include <zephyr/drivers/gpio.h>
-#include <zephyr/drivers/espi.h>
+#include <zephyr/drivers/espi/espi.h>
 #include <zephyr/pm/pm.h>
 #include <soc.h>
